Add support for Kubernetes 1.16.11, 1.17.7, 1.18.4 releases (#925)
* Add support for Kubernetes 1.16.11, 1.17.7, 1.18.4 releases

The latest releases are using CNI v0.8.6 which include a fix
for the CVE-2020-10749.

It's strongly advised to upgrade all your clusters
as soon as possible.

Ubuntu and CentOS:

The `kubernetes-cni` package has been marked as obsolete
and is not installed for clusters using the patched releases.
When upgrading to the patched release, the `kubernetes-cni` package
is removed automatically.

CoreOS and Flatcar:

All newly provisioned 1.16+ clusters use CNI v0.8.6.
The new version of CNI will be installed when upgrading
the cluster.

CentOS 7 known issue:

It's currently impossible to install versions older than
1.16.11, due to a problem with upstream packages.

* Update test fixtures