Making imagepull secrets configurable in KubeOne when using ContainerD
This commit extends the kubeone.yaml file by expanding the `registryConfiguration`
block to contain the `registryConfig` configuration bloci, which then is used to
configure the containerd daemon (in `/etc/containerd/config.tml`).
Specifically the `config.toml` file is extended to include the
`registry.mirrors."<regName>".auth` fields.

The containerd `config.toml` is written via a go-template in `pkg/scripts/render.go`
which in turn calls a go function `containerdCfg` which is implemented
in `pkg/scripts/funcs.go`.
The `containerdCfg` function is passed a second parameter: a list of
key-values, which then are de-serialised and written to the structs of
`containerdCRIRegistry` also located in `pkg/scripts/funcs.go`.
This struct is then serialised to toml, via
`"github.com/BurntSushi/toml"` just like before.

Please note: I did not back-port this change to the Docker-Shim
configuration, since it seemed a bit pointless to go the extra mile and
implement this config option in Docker as well, when Docker is
deprectated in the near future anyways.

Signed-off-by: Cedric Kienzler <github@cedric-kienzler.de>